Clean up long renderer JSX files
