+++ src/renderer/src/components/right-sidebar/HostedReviewActions.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines -- Why: Checks owns one compact hosted-review action bar; keeping GitHub PR and GitLab MR action branching together keeps provider parity visible. */
 import React, { useCallback, useMemo, useState } from 'react'
 import {
   LoaderCircle,
@@ -27,6 +26,7 @@ import type { PRInfo, Repo, Worktree } from '../../../../shared/types'
 import type { GitHubPRMergeMethod } from '../../../../shared/types'
 import { resolveGitHubPRMergeMethods } from '../../../../shared/github-pr-merge-methods'
 import { runWorktreeDelete } from '../sidebar/delete-worktree-flow'
+import { presentGitLabMRMergeState } from './gitlab-mr-merge-state'
 
 type HostedReviewActionInfo = Pick<
   HostedReviewInfo,
@@ -38,63 +38,6 @@ type HostedReviewActionInfo = Pick<
       'reviewDecision' | 'autoMergeEnabled' | 'mergeQueueRequired' | 'mergeStateStatus'
     >
   >
-
-function presentGitLabMRMergeState(review: HostedReviewActionInfo): {
-  label: string
-  tooltip: string
-  directMergeAvailable: boolean
-} {
-  if (review.state === 'merged') {
-    return {
-      label: 'Merged',
-      tooltip: 'This merge request is already merged',
-      directMergeAvailable: false
-    }
-  }
-  if (review.state === 'closed') {
-    return {
-      label: 'Closed',
-      tooltip: 'This merge request is closed',
-      directMergeAvailable: false
-    }
-  }
-  if (review.state === 'draft') {
-    return {
-      label: 'Draft',
-      tooltip: 'This merge request is still a draft',
-      directMergeAvailable: false
-    }
-  }
-  if (review.mergeable === 'CONFLICTING') {
-    return {
-      label: 'Conflicts',
-      tooltip: 'GitLab reports merge conflicts',
-      directMergeAvailable: false
-    }
-  }
-  if (review.status === 'failure') {
-    return {
-      label: 'Checks failed',
-      tooltip: 'GitLab says this MR can merge, but some pipeline jobs failed',
-      directMergeAvailable: true
-    }
-  }
-  if (review.status === 'pending') {
-    return {
-      label: 'Checks pending',
-      tooltip: 'GitLab says this MR can merge, but the pipeline is still running',
-      directMergeAvailable: true
-    }
-  }
-  return {
-    label: 'Able to merge',
-    tooltip:
-      review.mergeable === 'UNKNOWN'
-        ? 'GitLab has not reported a final merge status'
-        : 'GitLab says this MR can merge',
-    directMergeAvailable: true
-  }
-}
 
 export default function HostedReviewActions({
   review,
