+++ src/renderer/src/components/settings/NotificationSettingToggle.tsx
@@ -1,0 +1,48 @@
+import type { ReactNode } from 'react'
+import { Label } from '../ui/label'
+
+export type NotificationSettingToggleProps = {
+  label: string
+  description: string
+  checked: boolean
+  onToggle: () => void
+  disabled?: boolean
+  icon?: ReactNode
+}
+
+export function NotificationSettingToggle({
+  label,
+  description,
+  checked,
+  onToggle,
+  disabled = false,
+  icon
+}: NotificationSettingToggleProps): React.JSX.Element {
+  return (
+    <div className="flex items-center justify-between gap-4 py-2">
+      <div className="space-y-0.5">
+        <div className="flex items-center gap-2">
+          {icon}
+          <Label>{label}</Label>
+        </div>
+        <p className="text-xs text-muted-foreground">{description}</p>
+      </div>
+      <button
+        role="switch"
+        aria-checked={checked}
+        aria-label={label}
+        disabled={disabled}
+        onClick={onToggle}
+        className={`relative inline-flex h-5 w-9 shrink-0 items-center rounded-full border border-transparent transition-colors ${
+          checked ? 'bg-foreground' : 'bg-muted-foreground/30'
+        } ${disabled ? 'cursor-not-allowed opacity-50' : 'cursor-pointer'}`}
+      >
+        <span
+          className={`pointer-events-none block size-3.5 rounded-full bg-background shadow-sm transition-transform ${
+            checked ? 'translate-x-4' : 'translate-x-0.5'
+          }`}
+        />
+      </button>
+    </div>
+  )
+}
