+++ src/renderer/src/components/settings/TerminalAppearanceSection.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable max-lines -- Why: terminal visual controls stay together under Appearance so
-   search, previews, and the Ghostty import flow move as one user-facing surface. */
 import { useState } from 'react'
 import type { GlobalSettings } from '../../../../shared/types'
 import {
@@ -13,9 +11,7 @@ import {
   fontFamilyHasKnownLigatures,
   resolveTerminalLigaturesEnabled
 } from '../../../../shared/terminal-ligatures'
-import { Minus, Plus } from 'lucide-react'
 import { Button } from '../ui/button'
-import { Input } from '../ui/input'
 import {
   FontAutocomplete,
   NumberField,
@@ -40,6 +36,7 @@ import {
 import { DarkTerminalThemeSection, LightTerminalThemeSection } from './TerminalThemeSections'
 import { TerminalWindowSection } from './TerminalWindowSection'
 import { TerminalSettingsPreview } from './TerminalSettingsPreview'
+import { TerminalFontSizeSetting } from './TerminalFontSizeSetting'
 import { GhosttyImportModal } from './GhosttyImportModal'
 import type { UseGhosttyImportReturn } from './useGhosttyImport'
 import ghosttyIcon from '../../../../../resources/ghostty.svg'
@@ -88,56 +85,7 @@ export function TerminalAppearanceSection({
           </div>
 
           <div className="divide-y divide-border/40">
-            <SearchableSetting
-              title="Font Size"
-              description="Default terminal font size for new panes and live updates."
-              keywords={['terminal', 'typography', 'text size']}
-            >
-              <SettingsRow
-                label="Font Size"
-                description="Default terminal font size for new panes and live updates."
-                control={
-                  <div className="flex items-center gap-2">
-                    <Button
-                      variant="outline"
-                      size="icon-sm"
-                      onClick={() => {
-                        const next = Math.max(10, settings.terminalFontSize - 1)
-                        updateSettings({ terminalFontSize: next })
-                      }}
-                      disabled={settings.terminalFontSize <= 10}
-                    >
-                      <Minus className="size-3" />
-                    </Button>
-                    <Input
-                      type="number"
-                      min={10}
-                      max={24}
-                      value={settings.terminalFontSize}
-                      onChange={(e) => {
-                        const value = parseInt(e.target.value, 10)
-                        if (!Number.isNaN(value) && value >= 10 && value <= 24) {
-                          updateSettings({ terminalFontSize: value })
-                        }
-                      }}
-                      className="w-14 text-center tabular-nums"
-                    />
-                    <Button
-                      variant="outline"
-                      size="icon-sm"
-                      onClick={() => {
-                        const next = Math.min(24, settings.terminalFontSize + 1)
-                        updateSettings({ terminalFontSize: next })
-                      }}
-                      disabled={settings.terminalFontSize >= 24}
-                    >
-                      <Plus className="size-3" />
-                    </Button>
-                    <span className="text-xs text-muted-foreground">px</span>
-                  </div>
-                }
-              />
-            </SearchableSetting>
+            <TerminalFontSizeSetting settings={settings} updateSettings={updateSettings} />
 
             <SearchableSetting
               title="Font Family"
